Move episode feedkit to production
