+++ PodcastKo/PodcastKoCore/FeedKit/Get Episodes/FKEpisode.swift
@@ -1,0 +1,19 @@
+//
+//  FKEpisode.swift
+//  PodcastKoCore
+//
+//  Created by John Roque Jorillo on 5/30/21.
+//  Copyright © 2021 JohnRoque Inc. All rights reserved.
+//
+
+import Foundation
+
+public struct FKEpisode {
+    public let title: String?
+    public let pubDate: Date?
+    public let description: String?
+    public let author: String?
+    public let streamUrl: String?
+    
+    public let image: String?
+}
