+++ PodcastKo/PodcastKoCoreTests/FeedKit/Gateways/EpisodesFKGatewayTests.swift
@@ -9,43 +9,6 @@
 import XCTest
 import PodcastKoCore
 
-public protocol FeedKitClient {
-    typealias Result = Swift.Result<[FKEpisode], Error>
-    
-    func get(_ url: URL, completion: @escaping (Result) -> Void)
-}
-
-public struct FKEpisode {
-    public let title: String?
-    public let pubDate: Date?
-    public let description: String?
-    public let author: String?
-    public let streamUrl: String?
-    
-    public let image: String?
-}
-
-public final class EpisodesFKGateway: GetEpisodesUseCase {
-    
-    private let client: FeedKitClient
-    private let url: URL
-    
-    public enum Error: Swift.Error {
-        case connectivity
-    }
-    
-    public init(client: FeedKitClient, url: URL) {
-        self.client = client
-        self.url = url
-    }
-    
-    public func getEpisodes(completion: @escaping (GetEpisodesUseCase.Result) -> Void) {
-        client.get(self.url) { _ in
-            completion(.failure(Error.connectivity))
-        }
-    }
-}
-
 class EpisodesFKGatewayTests: XCTestCase {
     
     func test_init_doesNotRequestDataFromURL() {
@@ -77,10 +40,15 @@ class EpisodesFKGatewayTests: XCTestCase {
         let clientError = NSError(domain: "Test", code: 0)
         let (sut, client) = makeSUT()
         
-        expect(sut, toCompleteWith: failure(.connectivity)) {
+        expect(sut, toCompleteWith: .failure(clientError)) {
             client.complete(with: clientError)
         }
     }
+    
+//    func test_getEpisodes_deliversEpisodesOnClientResult() {
+//        let (sut, client) = makeSUT()
+//        
+//    }
     
     // MARK: - Helpers
     private func makeSUT(url: URL = anyURL(), file: StaticString = #filePath, line: UInt = #line) -> (sut: EpisodesFKGateway, client: FeedKitClientSpy) {
@@ -91,10 +59,6 @@ class EpisodesFKGatewayTests: XCTestCase {
         return (sut, client)
     }
     
-    private func failure(_ error: EpisodesFKGateway.Error) -> EpisodesFKGateway.Result {
-        EpisodesFKGateway.Result.failure(error)
-    }
-    
     private func expect(_ sut: EpisodesFKGateway, toCompleteWith expectedResult: EpisodesFKGateway.Result, when action: () -> Void, file: StaticString = #filePath, line: UInt = #line) {
         let exp = expectation(description: "Wait for load completion")
 
@@ -102,7 +66,7 @@ class EpisodesFKGatewayTests: XCTestCase {
             switch (receivedResult, expectedResult) {
             case let (.success(receivedItems), .success(expectedItems)):
                 XCTAssertEqual(receivedItems, expectedItems, file: file, line: line)
-            case let (.failure(receivedError as EpisodesFKGateway.Error), .failure(expectedError as EpisodesFKGateway.Error)):
+            case let (.failure(receivedError as NSError), .failure(expectedError as NSError)):
                 XCTAssertEqual(receivedError, expectedError, file: file, line: line)
             default:
                 XCTFail("Expected result \(expectedResult), got \(receivedResult) instead", file: file, line: line)
@@ -116,14 +80,14 @@ class EpisodesFKGatewayTests: XCTestCase {
         wait(for: [exp], timeout: 2.0)
     }
     
-    private class FeedKitClientSpy: FeedKitClient {
+    private class FeedKitClientSpy: EpisodeFeedKitClient {
         var requests: [URL] {
             messages.map { $0.url }
         }
         
-        private var messages = [(url: URL, completion: (FeedKitClient.Result) -> Void)]()
+        private var messages = [(url: URL, completion: (EpisodeFeedKitClient.Result) -> Void)]()
         
-        func get(_ url: URL, completion: @escaping (FeedKitClient.Result) -> Void) {
+        func get(_ url: URL, completion: @escaping (EpisodeFeedKitClient.Result) -> Void) {
             messages.append((url, completion))
         }
